nus-iss-fastapi/requirements.txt
fastapi 
uvicorn
multipart
pillow
scikit-learn==1.3.0
tensorflow==2.12.0






